requirements.txt
Django==1.8.6
djangorestframework==3.3.1
django-filter==0.11.0
psycopg2==2.6.1
wheel==0.24.0
Pillow==3.0.0